Added validation in HTML, JavaScript and Java in Maintenance Module

+++ src/main/java/com/cpi/is/service/impl/maintenance/DispatchTypeServiceImpl.java
@@ -35,14 +35,41 @@ public class DispatchTypeServiceImpl implements DispatchTypeService {
 
 	@Override
 	public String saveData(HttpServletRequest request) throws Exception {
-		return dispatchTypeDAO.saveData(
-				jsonToEntity(new JSONObject(request.getParameter("data"))));
+		String validation = validateData(request);
+		if(validation.equals("Success")) {
+			return dispatchTypeDAO.saveData(
+					jsonToEntity(new JSONObject(request.getParameter("data"))));
+		} else {
+			return validation;
+		}
 	}
 
 	@Override
 	public String deleteData(HttpServletRequest request) throws Exception {
-		return dispatchTypeDAO.deleteData(
-				jsonToEntity(new JSONObject(request.getParameter("data"))));
+		String validation = validateData(request);
+		if(validation.equals("Success")) {
+			return dispatchTypeDAO.deleteData(
+					jsonToEntity(new JSONObject(request.getParameter("data"))));
+		} else {
+			return "Unable to delete data";
+		}
+	}
+	
+	public String validateData(HttpServletRequest request) throws Exception{
+		JSONObject json = new JSONObject(request.getParameter("data"));
+		String validation = "Success";
+		
+		if (!json.has("dispatchTypeCd") || !(json.get("dispatchTypeCd") instanceof String)) {
+			validation = "Please fill-out the dispatch type form properly";
+		} else if(!json.has("dispatchTypeName") || !(json.get("dispatchTypeName") instanceof String)) {
+			validation = "Please fill-out the dispatch type form properly";
+		} else if (json.getString("dispatchTypeCd").length() < 1 || json.getString("dispatchTypeCd").length() > 50) {
+			validation = "Please fill-out the dispatch type form properly";
+		} else if (json.getString("dispatchTypeName").length() < 1 || json.getString("dispatchTypeName").length() > 200) {
+			validation = "Please fill-out the dispatch type form properly";
+		}
+		
+		return validation;
 	}
 
 }
